Get port id from the interface name. (#26)
* Get port id from the interface name.

In the CNI add API, the current implementation fetches
the port id of the interface from its name. It assumes the
name has the port id is part of the interface name as a
standard expression. This implementation is not reliable
in case the dpdk assigns a different port id for the interface.

Make it more dynamic and always fetch the port id by sending
a get request to the gNMI server.

Signed-off-by: Venkatesh Petla <neelakanta.venkatesh.petla@intel.com>

* Return error if failed to get port id.

Signed-off-by: Venkatesh Petla <neelakanta.venkatesh.petla@intel.com>

* Addressed review comments

Signed-off-by: Venkatesh Petla <neelakanta.venkatesh.petla@intel.com>

* Modified the way of populating the getGequest.

Signed-off-by: Venkatesh Petla <neelakanta.venkatesh.petla@intel.com>

* Get port id for SetupHostInterface from gNMIServer

Signed-off-by: Venkatesh Petla <neelakanta.venkatesh.petla@intel.com>

* Updated the p4info.txt to p4Info.txt

Signed-off-by: Venkatesh Petla <neelakanta.venkatesh.petla@intel.com>

* Change p4info to p4Info

Signed-off-by: Venkatesh Petla <neelakanta.venkatesh.petla@intel.com>

Signed-off-by: Venkatesh Petla <neelakanta.venkatesh.petla@intel.com>